(#237)refactor: 답변 service interface 구현했습니다.
